# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Papi Foundation API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /community/posts:
    get:
      operationId: listCommunityPosts
      tags: [community]
      summary: List community posts
      responses:
        "200":
          description: Community posts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/CommunityPost"
    post:
      operationId: createCommunityPost
      tags: [community]
      summary: Create a community post
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommunityPostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPost"
  /community/posts/{id}:
    delete:
      operationId: deleteCommunityPost
      tags: [community]
      summary: Delete a community post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /community/posts/{id}/reaction:
    put:
      operationId: reactToPost
      tags: [community]
      summary: Set or toggle a reaction on a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: Updated post
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CommunityPost"
  /community/posts/{id}/comments:
    get:
      operationId: listPostComments
      tags: [community]
      summary: List comments on a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Comments
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PostComment"
    post:
      operationId: createPostComment
      tags: [community]
      summary: Add a comment to a post
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CommentInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostComment"
  /community/comments/{id}:
    delete:
      operationId: deleteComment
      tags: [community]
      summary: Delete a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /community/comments/{id}/reaction:
    put:
      operationId: reactToComment
      tags: [community]
      summary: Set or toggle a reaction on a comment
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReactionInput"
      responses:
        "200":
          description: Updated comment
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PostComment"
...
components:
  schemas:
...
    ReactionCounts:
      type: object
      required: [like, love, celebrate, support, insightful]
      properties:
        like:
          type: integer
        love:
          type: integer
        celebrate:
          type: integer
        support:
          type: integer
        insightful:
          type: integer
    CommunityPost:
      type: object
      required: [id, authorId, authorName, content, createdAt, reactions, totalReactions, commentCount]
      properties:
        id:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        authorImageUrl:
          type: ["string", "null"]
        content:
          type: string
        imageUrl:
          type: ["string", "null"]
        createdAt:
          type: string
        reactions:
          $ref: "#/components/schemas/ReactionCounts"
        totalReactions:
          type: integer
        myReaction:
          type: ["string", "null"]
        commentCount:
          type: integer
    CommunityPostInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        imageUrl:
          type: string
    PostComment:
      type: object
      required: [id, postId, authorId, authorName, content, createdAt, reactions, totalReactions]
      properties:
        id:
          type: integer
        postId:
          type: integer
        authorId:
          type: string
        authorName:
          type: string
        authorImageUrl:
          type: ["string", "null"]
        content:
          type: string
        createdAt:
          type: string
        reactions:
          $ref: "#/components/schemas/ReactionCounts"
        totalReactions:
          type: integer
        myReaction:
          type: ["string", "null"]
    CommentInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
    ReactionInput:
      type: object
      required: [type]
      properties:
        type:
          type: string
          enum: [like, love, celebrate, support, insightful]
...